GravitationalPotentials

Documentation for GravitationalPotentials.

Index

GravitationalPotentials.EinastoDensity

GravitationalPotentials.MassDensityModel

GravitationalPotentials.PowerLawSphereDensity

GravitationalPotentials.SpiralGalaxyDensity

GravitationalPotentials.UniformCylinderDensity

GravitationalPotentials.UniformSphereDensity

GravitationalPotentials.mass_density

GravitationalPotentials.mass_density

GravitationalPotentials.mass_density

GravitationalPotentials.mass_density

GravitationalPotentials.mass_density

GravitationalPotentials.mass_density

Docstrings

GravitationalPotentials.MassDensityModel — Type

MassDensityModel

Abstract type for all mass density models. Any subtype of MassDensityModel should have a corresponding mass_density implementation.

GravitationalPotentials.EinastoDensity — Type

EinastoDensity <: MassDensityModel

Fields

central_density

harmonic_mean_radius

N

k

See Einasto, J. and Haud, U., “Galactic models with massive corona. I. Method”, Astronomy and Astrophysics, vol. 223, no. 1, pp. 89–94, 1989.

See also the implementation of mass_density for this type.

GravitationalPotentials.PowerLawSphereDensity — Type

PowerLawSphereDensity <: MassDensityModel

Model for a sphere with a power law density centered at the origin.

Fields

radius: radius of the sphere

scale_density: scale density of the sphere, density changes by scale_density^alpha at radius = scale_radius

scale_radius: scale radius of the sphere

alpha: power by which density reduces with respect to radius

See also the implementation of mass_density for this type.

GravitationalPotentials.SpiralGalaxyDensity — Type

SpiralGalaxyDensity <: MassDensityModel

Model for a spiral galaxy that has both a disk and a central bulge, with the bulge density dominating in the overlap.

The bulge is modeled as a uniform density sphere, and the disk is modeled as a uniform density cylinder.

Fields

bulge_radius: radius of the bulge

disk_radius: radius of the disk

disk_height: height of the disk

bulge_density: density of the bulge

disk_density: density of the cylinder

See also the implementation of mass_density for this type.

GravitationalPotentials.UniformCylinderDensity — Type

UniformCylinderDensity <: MassDensityModel

Model for a uniform density cylinder centered at the origin.

Fields

radius: radius of the cylinder

height: height of the cylinder

density: density of the cylinder

See also the implementation of mass_density for this type.

GravitationalPotentials.UniformSphereDensity — Type

UniformSphereDensity <: MassDensityModel

Model for a uniform density sphere centered at the origin.

Fields

radius: radius of the sphere

density: density of the sphere

See also the implementation of mass_density for this type.

GravitationalPotentials.mass_density — Function

mass_density(model::MassDensityModel, s, φ, z)

Given a mass density profile/model, return the corresponding mass density at cylindrical coordinates $(s, φ, z)$.

GravitationalPotentials.mass_density — Method

mass_density(model::EinastoDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = ρ₀ \exp\left[-(a / k a₀)^{1/N}\right]\]

where

$a = \sqrt{s^2 + z^2}$

$ρ₀$ = model.central_density

$a₀$ = model.harmonic_mean_radius

$k$ = model.k

$N$ = model.N

GravitationalPotentials.mass_density — Method

mass_density(model::PowerLawSphereDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_0 \left(\sqrt{s^2 + z^2}/R_0\right)^α & \text{if } \sqrt{s^2 + z^2} ≤ R_s \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_0$ = model.scale_density

$R_0$ = model.scale_radius

$α$ = model.alpha

$R_s$ = model.radius

GravitationalPotentials.mass_density — Method

mass_density(model::SpiralGalaxyDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_b & \text{if } \sqrt{s^2 + z^2} ≤ R_b \\ ρ_d & \text{if } s ≤ R_d, |z| ≤ H_d \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_b$ = model.bulge_density

$ρ_d$ = model.disk_density

$R_b$ = model.bulge_radius

$R_d$ = model.disk_radius

$H_d$ = model.disk_height

GravitationalPotentials.mass_density — Method

mass_density(model::UniformCylinderDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_0 & \text{if } s ≤ R_c, |z| ≤ H_c \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_0$ = model.density

$R_c$ = model.radius

$H_c$ = model.height

GravitationalPotentials.mass_density — Method

mass_density(model::UniformSphereDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_s & \text{if } \sqrt{s^2 + z^2} ≤ R_s \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_s$ = model.density

$R_s$ = model.radius