GravitationalPotentials

Documentation for GravitationalPotentials.

Index

GravitationalPotentials.EinastoDensity

GravitationalPotentials.MassDensityModel

GravitationalPotentials.PowerLawCylinderDensity

GravitationalPotentials.PowerLawSphereDensity

GravitationalPotentials.SpiralGalaxyDensity

GravitationalPotentials.UniformCylinderDensity

GravitationalPotentials.UniformSphereDensity

GravitationalPotentials.mass_density

GravitationalPotentials.mass_density

GravitationalPotentials.mass_density

GravitationalPotentials.mass_density

GravitationalPotentials.mass_density

GravitationalPotentials.mass_density

GravitationalPotentials.mass_density

Docstrings

GravitationalPotentials.MassDensityModel — Type

MassDensityModel

Abstract type for all mass density models. Any subtype of MassDensityModel should have a corresponding mass_density implementation.

GravitationalPotentials.EinastoDensity — Type

EinastoDensity <: MassDensityModel

Fields

ρ₀: central density

a₀: harmonic mean radius

N: structural parameter

k: normalizing constant

See Einasto, J. and Haud, U., “Galactic models with massive corona. I. Method”, Astronomy and Astrophysics, vol. 223, no. 1, pp. 89–94, 1989.

See also the implementation of mass_density for this type.

GravitationalPotentials.PowerLawCylinderDensity — Type

PowerLawCylinderDensity <: MassDensityModel

Model for a cylinder with a power law density centered at the origin.

Fields

r_c: radius of the cylinder

h_c: height of the cylinder

ρ₀: scale density of the sphere

r₀: scale radius of the cylinder

h₀: scale height of the cylinder

α: power by which density changes with respect to radius

β: power by which density changes with respect to radius

See also the implementation of mass_density for this type.

GravitationalPotentials.PowerLawSphereDensity — Type

PowerLawSphereDensity <: MassDensityModel

Model for a sphere with a power law density centered at the origin.

Fields

rₛ: radius of the sphere

ρ₀: scale density of the sphere, density changes by r₀^α at radius = r₀

r₀: scale radius of the sphere

α: power by which density reduces with respect to radius

See also the implementation of mass_density for this type.

GravitationalPotentials.SpiralGalaxyDensity — Type

SpiralGalaxyDensity <: MassDensityModel

Model for a spiral galaxy that has both a disk and a central bulge, with the bulge density dominating in the overlap.

The bulge is modeled as a uniform density sphere, and the disk is modeled as a uniform density cylinder.

Fields

r_bulge: radius of the bulge

r_disk: radius of the disk

h_disk: height of the disk

ρ_bulge: density of the bulge

ρ_disk: density of the cylinder

See also the implementation of mass_density for this type.

GravitationalPotentials.UniformCylinderDensity — Type

UniformCylinderDensity <: MassDensityModel

Model for a uniform density cylinder centered at the origin.

Fields

r_c: radius of the cylinder

h_c: height of the cylinder

ρ_c: density of the cylinder

See also the implementation of mass_density for this type.

GravitationalPotentials.UniformSphereDensity — Type

UniformSphereDensity <: MassDensityModel

Model for a uniform density sphere centered at the origin.

Fields

rₛ: radius of the sphere

ρₛ: density of the sphere

See also the implementation of mass_density for this type.

GravitationalPotentials.mass_density — Function

mass_density(model::MassDensityModel, s, φ, z)

Given a mass density profile/model, return the corresponding mass density at cylindrical coordinates $(s, φ, z)$.

GravitationalPotentials.mass_density — Method

mass_density(model::EinastoDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = ρ₀ \exp\left[-(a / k a₀)^{1/N}\right]\]

where

$a = √{s² + z²}$

$ρ₀$ = model.ρ₀

$a₀$ = model.a₀

$k$ = model.k

$N$ = model.N

GravitationalPotentials.mass_density — Method

mass_density(model::PowerLawCylinderDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_0 \left(s/R_0\right)^α \left(z/h_0\right)^β & \text{if } s ≤ R_c, \, z ≤ H_c \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_0$ = model.ρ₀

$R_0$ = model.r₀

$H_0$ = model.h₀

$α$ = model.α

$β$ = model.β

$R_c$ = model.r_c

$H_c$ = model.h_c

GravitationalPotentials.mass_density — Method

mass_density(model::PowerLawSphereDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_0 \left(\sqrt{s^2 + z^2}/R_0\right)^α & \text{if } \sqrt{s^2 + z^2} ≤ R_s \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_0$ = model.ρ₀

$R_0$ = model.r₀

$α$ = model.α

$R_s$ = model.rₛ

GravitationalPotentials.mass_density — Method

mass_density(model::SpiralGalaxyDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_b & \text{if } \sqrt{s^2 + z^2} ≤ R_b \\ ρ_d & \text{if } s ≤ R_d, |z| ≤ H_d \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_b$ = model.ρ_bulge

$ρ_d$ = model.ρ_disk

$R_b$ = model.r_bulge

$R_d$ = model.r_disk

$H_d$ = model.h_disk

GravitationalPotentials.mass_density — Method

mass_density(model::UniformCylinderDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_0 & \text{if } s ≤ R_c, |z| ≤ H_c \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_0$ = model.ρ_c

$R_c$ = model.r_c

$H_c$ = model.h_c

GravitationalPotentials.mass_density — Method

mass_density(model::UniformSphereDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_s & \text{if } \sqrt{s^2 + z^2} ≤ R_s \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_s$ = model.ρₛ

$R_s$ = model.rₛ