API reference

Documentation for GravitationalPotentials.

Density models

GravitationalPotentials.MassDensityModel — Type

abstract type MassDensityModel

Abstract type for all mass density models. Any subtype of MassDensityModel should have a corresponding mass_density implementation.

GravitationalPotentials.mass_density — Function

mass_density(model::MassDensityModel, s, φ, z)

Given a mass density profile/model, return the corresponding mass density at cylindrical coordinates $(s, φ, z)$.

Uniform sphere model

GravitationalPotentials.UniformSphereDensity — Type

struct UniformSphereDensity{L, MD} <: MassDensityModel

Model for a uniform density sphere centered at the origin.

Fields

rₛ::Any: radius of the sphere

ρₛ::Any: density of the sphere

See also the implementation of mass_density for this type.

GravitationalPotentials.mass_density — Method

mass_density(model::UniformSphereDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_s & \text{if } \sqrt{s^2 + z^2} ≤ R_s \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_s$ = model.ρₛ

$R_s$ = model.rₛ

Power law sphere model

GravitationalPotentials.PowerLawSphereDensity — Type

struct PowerLawSphereDensity{L, MD} <: MassDensityModel

Model for a sphere with a power law density centered at the origin.

Fields

rₛ::Any: radius of the sphere

ρ₀::Any: scale density of the sphere, density changes by r₀^α at radius = r₀

r₀::Any: scale radius of the sphere

α::Float64: power by which density reduces with respect to radius

See also the implementation of mass_density for this type.

GravitationalPotentials.mass_density — Method

mass_density(model::PowerLawSphereDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_0 \left(\sqrt{s^2 + z^2}/R_0\right)^α & \text{if } \sqrt{s^2 + z^2} ≤ R_s \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_0$ = model.ρ₀

$R_0$ = model.r₀

$α$ = model.α

$R_s$ = model.rₛ

Uniform cylinder density model

GravitationalPotentials.UniformCylinderDensity — Type

struct UniformCylinderDensity{L, MD} <: MassDensityModel

Model for a uniform density cylinder centered at the origin.

Fields

r_c::Any: radius of the cylinder

h_c::Any: height of the cylinder

ρ_c::Any: density of the cylinder

See also the implementation of mass_density for this type.

GravitationalPotentials.mass_density — Method

mass_density(model::UniformCylinderDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_0 & \text{if } s ≤ R_c, |z| ≤ H_c \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_0$ = model.ρ_c

$R_c$ = model.r_c

$H_c$ = model.h_c

GravitationalPotentials.onaxispotential — Method

onaxispotential(model::UniformCylinderDensity, z)

Potential for test point at (0, φ, z). Technically φ is not well-defined for s=0, but it's irrelevant here.

Power law cylinder density model

GravitationalPotentials.PowerLawCylinderDensity — Type

struct PowerLawCylinderDensity{L, MD} <: MassDensityModel

Model for a cylinder with a power law density centered at the origin.

Fields

r_c::Any: radius of the cylinder

h_c::Any: height of the cylinder

ρ₀::Any: scale density of the sphere

r₀::Any: scale radius of the cylinder

h₀::Any: scale height of the cylinder

α::Float64: power by which density changes with respect to radius

β::Float64: power by which density changes with respect to height

See also the implementation of mass_density for this type.

GravitationalPotentials.mass_density — Method

mass_density(model::PowerLawCylinderDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_0 \left(s/R_0\right)^α \left(z/h_0\right)^β & \text{if } s ≤ R_c, \, z ≤ H_c \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_0$ = model.ρ₀

$R_0$ = model.r₀

$H_0$ = model.h₀

$α$ = model.α

$β$ = model.β

$R_c$ = model.r_c

$H_c$ = model.h_c

Spiral galaxy model

GravitationalPotentials.SpiralGalaxyDensity — Type

struct SpiralGalaxyDensity{L, MD} <: MassDensityModel

Model for a spiral galaxy that has both a disk and a central bulge, with the bulge density dominating in the overlap.

The bulge is modeled as a uniform density sphere, and the disk is modeled as a uniform density cylinder.

Fields

r_bulge::Any: radius of the bulge

r_disk::Any: radius of the disk

h_disk::Any: height of the disk

ρ_bulge::Any: density of the bulge

ρ_disk::Any: density of the disk

See also the implementation of mass_density for this type.

GravitationalPotentials.mass_density — Method

mass_density(model::SpiralGalaxyDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \begin{cases} ρ_b & \text{if } \sqrt{s^2 + z^2} ≤ R_b \\ ρ_d & \text{if } s ≤ R_d, |z| ≤ H_d \\ 0 & \text{otherwise} \end{cases}\]

where

$ρ_b$ = model.ρ_bulge

$ρ_d$ = model.ρ_disk

$R_b$ = model.r_bulge

$R_d$ = model.r_disk

$H_d$ = model.h_disk

Einasto density model

GravitationalPotentials.EinastoDensity — Type

struct EinastoDensity{L, MD} <: MassDensityModel

Fields

ρ₀::Any: central density

a₀::Any: harmonic mean radius

N::Float64: structural parameter

k::Float64: normalizing constant

See Einasto, J. and Haud, U., “Galactic models with massive corona. I. Method”, Astronomy and Astrophysics, vol. 223, no. 1, pp. 89–94, 1989.

See also the implementation of mass_density for this type.

GravitationalPotentials.mass_density — Method

mass_density(model::EinastoDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = ρ₀ \exp\left[-\left(\frac{a}{k a₀}\right)^{1/N}\right]\]

where

$a = \sqrt{s² + z²}$

$ρ₀$ = model.ρ₀

$a₀$ = model.a₀

$k$ = model.k

$N$ = model.N

Navarro–Frenk–White Profile

GravitationalPotentials.NFWDensity — Type

struct NFWDensity{L, MD} <: MassDensityModel

Fields

ρ₀::Any: Scale density

r₀::Any: Scale radius

GravitationalPotentials.mass_density — Method

mass_density(model::NFWDensity, s, φ, z)

Density model

\[ρ(s, φ, z) = \frac{ρ₀}{\frac{r}{r_0} \left(1 + \frac{r}{r_0}\right)^2}\]

where $r = \sqrt{s^2 + z^2}$